medicalpdf/requirements.txt
Flask
Flask-CORS
pytesseract
pdf2image
Pillow
python-dotenv
PyPDF2
